test(toolkit): address PR review — relocate ECDSA wallet tests, make maintenance key fallible
- Move the ledger-9-only ECDSA `UnshieldedWallet` tests out of the shared
  `common` module (compiled once per generation) into a `ledger_9`-only
  `ecdsa_wallet_tests` module, so they no longer report a misleading
  `ledger_7::…::ecdsa_… ok` and drop the runtime generation guard.
- `UnshieldedWallet::maintenance_verifying_key` now returns `Option`, so an
  address-only wallet no longer panics; callers make the invariant explicit.
- Randomize the deploy rng-seed in `ecdsa_contract_committees_e2e` so re-runs
  against a persistent node don't collide on the contract address.
- Add `ecdsa_contract_committees_e2e` to the serial `contract-e2e` nextest group.
